tools/requirements.txt
# Python Tools Requirements
# OmniScript helper tools dependencies

# Core utilities
requests>=2.28.0
pyyaml>=6.0
rich>=13.0.0

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0

# Linting
ruff>=0.1.0
mypy>=1.0.0

# Docker SDK (optional for advanced operations)
docker>=6.0.0
